Add test for SOURCE_DATE_EPOCH
